added test router to explore the Web-interface
